+++ example/lib/main.dart
@@ -20,13 +20,13 @@ class MyApp extends StatelessWidget {
   }
 }
 
-class MyHomePage extends BlocConsumerWidget<CalculatorBloc, BlocState> {
-  const MyHomePage({super.key, required this.title});
+class MyConsumerHomePage extends BlocConsumerWidget<CalculatorBloc, BlocState> {
+  const MyConsumerHomePage({super.key, required this.title});
 
   final String title;
 
   @override
-  CalculatorBloc? bloc(BuildContext context) => CalculatorBloc();
+  CalculatorBloc bloc(BuildContext context) => CalculatorBloc();
 
   @override
   void onMount(CalculatorBloc bloc) => bloc.add(CalculatorEvent.onMount);
@@ -48,7 +48,58 @@ class MyHomePage extends BlocConsumerWidget<CalculatorBloc, BlocState> {
   }
 
   @override
-  Widget buildWithState(context, bloc, state) {
+  Widget build(context, bloc, state) {
+    final textTheme = Theme.of(context).textTheme;
+    return Scaffold(
+      appBar: AppBar(title: Text(title)),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text('Counter value:'),
+            Text('${state.counter}', style: textTheme.headlineMedium),
+          ],
+        ),
+      ),
+      bottomNavigationBar: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            FloatingActionButton(
+              onPressed: () => bloc.add(CalculatorEvent.remove),
+              tooltip: 'Decrement',
+              child: const Icon(Icons.exposure_minus_1),
+            ),
+            TextButton(
+              onPressed: () => bloc.add(CalculatorEvent.clear),
+              child: const Text('Clear counter'),
+            ),
+            FloatingActionButton(
+              onPressed: () => bloc.add(CalculatorEvent.add),
+              tooltip: 'Increment',
+              child: const Icon(Icons.exposure_plus_1),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class MyBuilderHomePage extends BlocBuilderWidget<CalculatorBloc, BlocState> {
+  const MyBuilderHomePage({super.key, required this.title});
+
+  final String title;
+
+  @override
+  CalculatorBloc bloc(BuildContext context) => CalculatorBloc();
+
+  @override
+  void onMount(CalculatorBloc bloc) => bloc.add(CalculatorEvent.onMount);
+
+  @override
+  Widget build(context, bloc, state) {
     final textTheme = Theme.of(context).textTheme;
     return Scaffold(
       appBar: AppBar(title: Text(title)),
@@ -94,16 +145,16 @@ class MySelectorHomePage
   final String title;
 
   @override
-  CalculatorCubit? bloc(BuildContext context) => CalculatorCubit();
+  CalculatorCubit bloc(BuildContext context) => CalculatorCubit();
 
   @override
   void onMount(CalculatorCubit bloc) => bloc.changeCounter(-1);
 
   @override
-  int selector(CalculatorCubit cubit, CubitState state) => state.counter;
+  int selector(CubitState state) => state.counter;
 
   @override
-  Widget buildWithState(context, bloc, selector) {
+  Widget build(context, bloc, selector) {
     final textTheme = Theme.of(context).textTheme;
     return Scaffold(
       appBar: AppBar(title: Text(title)),
@@ -138,6 +189,34 @@ class MySelectorHomePage
           ],
         ),
       ),
+    );
+  }
+}
+
+class MyListenerHomePage
+    extends BlocListenerWidget<CalculatorCubit, CubitState> {
+  const MyListenerHomePage({super.key, required this.title});
+
+  final String title;
+
+  @override
+  CalculatorCubit bloc(BuildContext context) => CalculatorCubit();
+
+  @override
+  void onMount(CalculatorCubit bloc) => bloc.changeCounter(-1);
+
+  @override
+  void listener(BuildContext context, CalculatorCubit bloc, CubitState state) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(content: Text('Counter value: ${state.counter}')),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context, CalculatorCubit bloc) {
+    return Scaffold(
+      appBar: AppBar(title: Text(title)),
+      body: const Center(child: Text('Click to show SnackBar with value')),
     );
   }
 }