+++ lib/bloc_widgets.dart
@@ -1,5 +1,3 @@
 library bloc_widgets;
 
-export 'src/bloc_builder_widget.dart';
-export 'src/bloc_consumer_widget.dart';
-export 'src/bloc_selector_widget.dart';
+export 'src/bloc_widget.dart';