+++ lib/src/bloc_listener_widget.dart
@@ -1,20 +1,18 @@
 part of 'bloc_widget.dart';
 
-abstract class BlocConsumerWidget<B extends BlocBase<S>, S>
+abstract class BlocListenerWidget<B extends BlocBase<S>, S>
     extends _BlocWidget {
-  const BlocConsumerWidget({super.key});
+  const BlocListenerWidget({super.key});
 
   bool listenWhen(S prev, S next) => true;
 
   void listener(BuildContext context, B bloc, S state) {}
 
-  bool buildWhen(S prev, S next) => true;
-
   void onMount(B bloc) {}
 
   B bloc(BuildContext context) => context.read<B>();
 
-  Widget build(BuildContext context, B bloc, S state);
+  Widget build(BuildContext context, B bloc);
 
   @override
   Widget _build(BuildContext context) {
@@ -22,12 +20,11 @@ abstract class BlocConsumerWidget<B extends BlocBase<S>, S>
 
     onMount(cubit);
 
-    return BlocConsumer<B, S>(
+    return BlocListener<B, S>(
       bloc: cubit,
-      buildWhen: buildWhen,
-      builder: (context, state) => build(context, cubit, state),
       listenWhen: listenWhen,
       listener: (context, state) => listener(context, cubit, state),
+      child: build(context, cubit),
     );
   }
 }