+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+* **Breaking change:** method `buildWithState` replaced with more conventional `build` method.
+* Added  `BlocListenerWidget` widgets.
+* `flutter_bloc` dependency updated.
+
 ## 1.0.0
 
 * First release of `bloc_widgets` package.
+